requirements.txt
Flask==3.0.0
flask-session==0.5.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
